swithcing computers

+++ 16-the-wild-oasis/src/services/apiCabins.js
@@ -12,35 +12,49 @@ export async function getCabins() {
   return data;
 }
 
-export async function createCabin(newCabin) {
-  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
+  const imageName = `${crypto.randomUUID()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  // 1. Create cabin
-  const { data, error } = await supabase
-    .from("Cabins")
-    .insert([{ ...newCabin, image: imagePath }]);
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
+
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
   // .select();
 
   if (error) {
-    console.log(error);
-    throw new Error("Cabin could not be created");
+    console.log(error.message);
+    throw new Error(error.message || "Cabin could not be created");
   }
 
   // 2. Upload image
-  const { error: storageError } = await supabase.storage
-    .from("cabin-images")
-    .upload(imageName, newCabin.image);
+  if (!hasImagePath) {
+    const { error: storageError } = await supabase.storage
+      .from("cabin-images")
+      .upload(imageName, newCabin.image);
 
-  // 3. Delete the camin if there was an error uploading image
-  if (storageError) {
-    await supabase.from("cabins").delete().eq("id", data.id);
+    // 3. Delete the camin if there was an error uploading image
+    if (storageError) {
+      await supabase.from("cabins").delete().eq("id", data.id);
 
-    console.error(storageError);
-    throw new Error('Cabin image could not be uploaded and the cabin was not created');
+      console.error(storageError);
+      throw new Error(
+        "Cabin image could not be uploaded and the cabin was not created"
+      );
+    }
   }
 
   return data;
